==> notebooks/04.2_JN_bloom_metabase_updated.ipynb <==
#import data processing libraries

import os
import pandas as pd
import numpy as np
import math as math
import datetime as dt
from scipy.stats import mode
import psycopg2

#import visualization libraries
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
%matplotlib inline

pd.set_option('display.max_colwidth', 500)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 300)

#suppress scientific notation to 2 decimal places

pd.set_option('display.float_format', lambda x: '%.2f' % x)

host = '157.245.248.249'
port = int(5432)
dbname = 'ubuntu'
user = 'jacklinengenia'
password = 'x3MX&8#!'


def get_query_results_postgres():
    with psycopg2.connect(host = host,
                          port = port,
                          database = dbname,
                          user = user,
                          password = password) as conn:
        sql = "SELECT store_number, phone, transaction_id, amount, transaction_time FROM bloomlive.till_activity_materialized_view WHERE type = 'c2b' AND transaction_time <= '2022-09-22 23:59:59' AND transaction_time >= '2022-08-22 00:00:00'" 
        df = pd.read_sql(sql, conn)
        
    return df

    conn.close()
    
if __name__ == "__main__":
    get_query_results_postgres()  

df1 = get_query_results_postgres()

df1.head()

#trim data to only have target colums
target_columns = ["store_number","phone","transaction_id","amount","transaction_time"]

df1 = df1[target_columns]

#convert date columns
datetime_cols = ["transaction_time"]

df1[datetime_cols] = df1[datetime_cols].apply(pd.to_datetime, errors='coerce')


#convert specific int/float columns to string type
int_to_string_cols = ["store_number"]

df1[int_to_string_cols] = df1[int_to_string_cols].astype(str)

#drop any duplicate transactions
df1 = df1.loc[~df1["transaction_id"].duplicated()]


#drop any rows with national id being blank
df1 = df1.loc[df1["store_number"].notnull()]

# strip transaction time column to only include Y-m-d ==> this converts datetime column to string
df1["transaction_time"] = df1["transaction_time"].apply(lambda x: x.strftime("%Y-%m-%d")) 


# convert transaction time column to datetime
datetime_cols = ["transaction_time"]
df1[datetime_cols] = df1[datetime_cols].apply(pd.to_datetime, errors='coerce')


df1.head(3)

df1['transaction_time'].max()

host = '157.245.248.249'
port = int(5432)
dbname = 'ubuntu'
user = 'jacklinengenia'
password = 'x3MX&8#!'


def get_query_results_postgres():
    with psycopg2.connect(host = host,
                          port = port,
                          database = dbname,
                          user = user,
                          password = password) as conn:
        sql = "SELECT store_number, phone, transaction_id, amount, transaction_time FROM bloomlive.till_activity_materialized_view WHERE type = 'c2b' AND transaction_time <= '2022-08-21 23:59:59' AND transaction_time >= '2022-07-22 00:00:00'" 
        df = pd.read_sql(sql, conn)
        
    return df

    conn.close()
    
if __name__ == "__main__":
    get_query_results_postgres()  

df2 = get_query_results_postgres()

df2.head(3)

#trim data to only have target colums
target_columns = ["store_number","phone","transaction_id","amount","transaction_time"]

df2 = df2[target_columns]

#convert date columns
datetime_cols = ["transaction_time"]

df2[datetime_cols] = df2[datetime_cols].apply(pd.to_datetime, errors='coerce')


#convert specific int/float columns to string type
int_to_string_cols = ["store_number"]

df2[int_to_string_cols] = df2[int_to_string_cols].astype(str)

#drop any duplicate transactions
df2 = df2.loc[~df2["transaction_id"].duplicated()]


#drop any rows with national id being blank
df2 = df2.loc[df2["store_number"].notnull()]

# strip transaction time column to only include Y-m-d ==> this converts datetime column to string
df2["transaction_time"] = df2["transaction_time"].apply(lambda x: x.strftime("%Y-%m-%d")) 


# convert transaction time column to datetime
datetime_cols = ["transaction_time"]
df2[datetime_cols] = df2[datetime_cols].apply(pd.to_datetime, errors='coerce')


df2.head(3)

host = '157.245.248.249'
port = int(5432)
dbname = 'ubuntu'
user = 'jacklinengenia'
password = 'x3MX&8#!'


def get_query_results_postgres():
    with psycopg2.connect(host = host,
                          port = port,
                          database = dbname,
                          user = user,
                          password = password) as conn:
        sql = "SELECT store_number, phone, transaction_id, amount, transaction_time FROM bloomlive.till_activity_materialized_view WHERE type = 'c2b' AND transaction_time <= '2022-07-21 23:59:59' AND transaction_time >= '2022-06-22 00:00:00'" 
        df = pd.read_sql(sql, conn)
        
    return df

    conn.close()
    
if __name__ == "__main__":
    get_query_results_postgres()  

df3 = get_query_results_postgres()

df3.head(3)

#trim data to only have target colums
target_columns = ["store_number","phone","transaction_id","amount","transaction_time"]

df3 = df3[target_columns]

#convert date columns
datetime_cols = ["transaction_time"]

df3[datetime_cols] = df3[datetime_cols].apply(pd.to_datetime, errors='coerce')


#convert specific int/float columns to string type
int_to_string_cols = ["store_number"]

df3[int_to_string_cols] = df3[int_to_string_cols].astype(str)

#drop any duplicate transactions
df3 = df3.loc[~df3["transaction_id"].duplicated()]


#drop any rows with national id being blank
df3 = df3.loc[df3["store_number"].notnull()]

# strip transaction time column to only include Y-m-d ==> this converts datetime column to string
df3["transaction_time"] = df3["transaction_time"].apply(lambda x: x.strftime("%Y-%m-%d")) 


# convert transaction time column to datetime
datetime_cols = ["transaction_time"]
df3[datetime_cols] = df3[datetime_cols].apply(pd.to_datetime, errors='coerce')


df3.head(3)

df = pd.concat([df1, df2, df3], ignore_index=True)

df.head()

#trim data to only have target colums
target_columns = ["store_number","phone","transaction_id","amount","transaction_time"]

df = df[target_columns]

#convert date columns
datetime_cols = ["transaction_time"]

df[datetime_cols] = df[datetime_cols].apply(pd.to_datetime, errors='coerce')


#convert specific int/float columns to string type
int_to_string_cols = ["store_number"]

df[int_to_string_cols] = df[int_to_string_cols].astype(str)

#drop any duplicate transactions
df = df.loc[~df["transaction_id"].duplicated()]


#drop any rows with national id being blank
df = df.loc[df["store_number"].notnull()]

# strip transaction time column to only include Y-m-d ==> this converts datetime column to string
df["transaction_time"] = df["transaction_time"].apply(lambda x: x.strftime("%Y-%m-%d")) 


# convert transaction time column to datetime
datetime_cols = ["transaction_time"]
df[datetime_cols] = df[datetime_cols].apply(pd.to_datetime, errors='coerce')

df.head(3)

df['store_number'].nunique()

df.shape

df1['transaction_time'].max()

df['transaction_time'].min()

#### Generate aggregate summaries

agg_summary = df1.groupby("store_number").agg({
    "amount":"sum",
    "transaction_time":"min"
}).reset_index()

agg_summary.rename(
    columns={
        "amount": "approx_30_days_trx_val",
        "transaction_time": "most_recent_trx_date_past_30_days",
    }, inplace=True)

agg_summary.head()

#aggregate for latest trx date

agg_summary = pd.merge(agg_summary, (df1.groupby("store_number")["transaction_time"].max().rename("last_trx_date").reset_index()), on="store_number")

#calculate expected trx days

agg_summary["expected_trx_days"] = ((agg_summary["last_trx_date"]-agg_summary["most_recent_trx_date_past_30_days"]).dt.days)+1

#get unique trx days

agg_summary = pd.merge(agg_summary, (df1.groupby("store_number")["transaction_time"].nunique().rename("actual_trx_days").reset_index()), on="store_number")

#calculate consistency

agg_summary["page_active_days"] = round(agg_summary["actual_trx_days"]/agg_summary["expected_trx_days"], 2)

agg_summary.head()

#load df with store number inferences

#inference_df_path = "C:\\Project_summaries\\Bloom\\Bloom all_loans\\20220726\\Analysis_summaries\\"
inference_df = pd.read_excel("/content/Bloom_clients_inference_summary_20220922.xlsx")
inference_df.drop(columns="Unnamed: 0", inplace=True)
inference_df["store_number"] = inference_df["store_number"].astype("str")

#merge df to add inference column

agg_summary = agg_summary.merge(inference_df, how="left", on="store_number")

#fill rest of missing values with zeros

cols_fillna = ["inference_col"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    agg_summary[col].fillna("No_rules_relaxed",inplace=True)

#get num days since last trx

#today = dt.datetime(2022,2,26)
#today = pd.Timestamp(today)
today = (pd.Timestamp.today()).strftime('%Y-%m-%d')
today = pd.Timestamp(today)

#not counting the end date because the scoring refresh might be triggered early in the morning or before working day is over
agg_summary["days_since_last_trx"] = (today - agg_summary["last_trx_date"]).dt.days

def recency_check(df):
    """
    Function to assess whether store number has recently been trading up to a certain allowed threshold i.e. 5 days for\
    those that don't qualify for limit stabilization and upto 7 days for those that qualify
    
    Inputs:   
    1) The inference column indicating whether rules are to be relaxed OR not,
    2) Recency tracking column i.e. num of days since store number last had a transaction
    3) Recency thresholds
    
    Outputs:
    A column denoting a boolean yes OR no wrt to whether a customer met the required recency threshold
    """   
    inference_col = df["inference_col"]
    inference_col_target = "relax_rules"
    recency_col = df["days_since_last_trx"]
    no_rules_relaxed_recency_threshold = 5
    rules_relaxed_recency_threshold = 7
    
    #choice responses
    transaction_boolean_accepted = "Yes"
    transaction_boolean_rejected = "No"
    
    
    conditions = [
        recency_col.le(no_rules_relaxed_recency_threshold),
        inference_col.str.match(inference_col_target) & recency_col.le(rules_relaxed_recency_threshold),
        recency_col.gt(no_rules_relaxed_recency_threshold),  
    ]
    
    choices = [
        transaction_boolean_accepted,
        transaction_boolean_accepted,
        transaction_boolean_rejected,
    ]
    
    new_col = np.select(conditions, choices)
    
    return new_col

#apply the function to the df to create 21 day adjusted product limit allocation
agg_summary["transacted_last_5_days"] = recency_check(agg_summary)

def weight_till_recency(df):
    """
    Function to assess the weight to be assigned based on till recency for customers who qualify for limit stabilization
    
    Inputs:   
    1) The inference column indicating whether rules are to be relaxed OR not,
    2) Recency tracking column i.e. num of days since store number last had a transaction
    
    Outputs:
    A column assigning the assigned weight for till recency
    """       
    recency = df["days_since_last_trx"]
    inference_col = df["inference_col"]
    inference_col_target = "relax_rules"
    
    conditions = [
        recency.le(5),
        inference_col.str.match(inference_col_target) & recency.le(5),
        inference_col.str.match(inference_col_target) & recency.gt(5) & recency.le(6),
        inference_col.str.match(inference_col_target) & recency.gt(6) & recency.le(7),
        recency.gt(7),
    ]
    
    choices = [
        1,
        1,
        0.9,
        0.7,
        0 
    ]
    
    weight_till_recency_col = np.select(conditions, choices)
    
    return weight_till_recency_col

#apply function
agg_summary["weight_till_recency"] = weight_till_recency(agg_summary)

agg_summary.to_excel("C:\\Project_summaries\\Bloom\\Bloom all_loans\\20220726\Till_data_summaries\\agg_summary.xlsx")

host = '157.245.248.249'
port = int(5432)
dbname = 'ubuntu'
user = 'jacklinengenia'
password = 'x3MX&8#!'


def get_query_results_postgres():
    with psycopg2.connect(host = host,
                          port = port,
                          database = dbname,
                          user = user,
                          password = password) as conn:
        sql = "select * from bloomlive.client_summary_view csv2 where is_iprs_validated is true"
        df = pd.read_sql(sql, conn)
        
    return df

    conn.close()

# loading data from scoring results table
customer_details = get_query_results_postgres()

customer_details.head(3)

customer_details = customer_details.drop_duplicates(subset=['store_number'], keep='last')

customer_details.shape

#load customer_details with store number & national id mapping

#customer_details = pd.read_csv(customers_path+"iprs_report_2021-10-14.csv")

#replace whitespaces on column headers with 
customer_details.columns = customer_details.columns.str.replace(' ','_')

#rename Pandas columns to lower case
customer_details.columns = customer_details.columns.str.lower()

#rename column to make it clearer
#customer_details.rename(columns = {"customeridnumber": "national_id", "accountnumber":"store_number"}, inplace = True)

#trim customer_details to only have target columns
customer_target_columns = ["store_number", "national_id", 'mobile_number']

customer_details = customer_details[customer_target_columns]

#convert specific int/float columns to string type
int_to_string_cols_2 = ["store_number", "national_id", "mobile_number"]

customer_details[int_to_string_cols_2] = customer_details[int_to_string_cols_2].astype(str)

#merge dfs to map store numbers to national ids

agg_summary = agg_summary.merge(customer_details, how="left", on="store_number")

#rearrange column order

agg_summary_cols = agg_summary.columns.to_list()

agg_summary_cols = agg_summary_cols[-1:] + agg_summary_cols[:-1]

agg_summary = agg_summary[agg_summary_cols]

agg_summary.to_excel("Bloom_scoring_trx_data_20220922.xlsx")
