==> notebooks/15.1_JN_adding_merchants_qualifying_21_day_graduation_original.ipynb <==
#import data processing libraries

import os
import pandas as pd
import numpy as np
import math as math
import datetime as dt
from scipy.stats import mode
import psycopg2 #reading data from Postgres DB
import sys
import openpyxl

pd.set_option('display.max_colwidth', 500)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 300)

#suppress scientific notation to 2 decimal places
#pd.set_option('display.float_format', lambda x: '%.2f' % x)

#import visualization libraries

import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
%matplotlib inline

# Connection parameters
param_dic = {
    "host"      : "157.245.248.249",
    "database"  : "ubuntu",
    "user"      : "jacklinengenia",
    "password"  : "x3MX&8#!"
}

def connect(params_dic):
    """ Connect to the PostgreSQL database server """
    conn = None
    try:
        # connect to the PostgreSQL server
        print('Connecting to the PostgreSQL database...')
        conn = psycopg2.connect(**params_dic)
    except (Exception, psycopg2.DatabaseError) as error:
        print(error)
        sys.exit(1) 
    print("Connection successful")
    return conn

def postgresql_to_dataframe(conn, select_query, column_names):
    """
    Tranform a SELECT query into a pandas dataframe
    """
    cursor = conn.cursor()
    try:
        cursor.execute(select_query)
    except (Exception, psycopg2.DatabaseError) as error:
        print("Error: %s" % error)
        cursor.close()
        return 1
    
    # Naturally we get a list of tupples
    tupples = cursor.fetchall()
    cursor.close()
    
    # Convert the list of tuples in to a pandas dataframe
    df = pd.DataFrame(tupples, columns=column_names)
    return df

# Connect to the database and load the scoring data bloomlive table
conn = connect(param_dic)

columns = "client_mifos_id,client_mobile_number,loan_status,loan_mifos_id,term_frequency,principal_disbursed,\
principal_repaid,interest_charged,interest_repaid,\
fee_charges_charged,fee_charges_repaid,\
penalty_charges_charged,penalty_charges_repaid,\
total_outstanding,\
disbursed_on_date,expected_matured_on_date,\
closed_on_date,store_number, bloom_version"

column_names = columns.strip().split(",")
# Execute the "SELECT cols" query
df = postgresql_to_dataframe(conn,
                             "select \
                             client_mifos_id,client_mobile_number,loan_status,loan_mifos_id,term_frequency,principal_disbursed,\
                             principal_repaid,interest_charged,interest_repaid,\
                             fee_charges_charged,fee_charges_repaid,\
                             penalty_charges_charged,penalty_charges_repaid,\
                             total_outstanding,\
                             disbursed_on_date,expected_matured_on_date,\
                             closed_on_date,store_number, bloom_version\
                             from bloomlive.loans_fact_table_materialized_summary_view",
                             column_names)

#replace None type datatype with Nan values across entire df

df = df.fillna(value=np.nan)

#convert date columns
datetime_cols = ["disbursed_on_date", "expected_matured_on_date", "closed_on_date"]

df[datetime_cols] = df[datetime_cols].apply(pd.to_datetime, errors='coerce')


#convert specific int/float columns to string type
int_to_string_cols = ["client_mifos_id", "client_mobile_number"]

df[int_to_string_cols] = df[int_to_string_cols].astype(str)


#convert specific string columns to float, first by leaving out the non-targeted columns
string_to_float_cols = df.columns.drop(["client_mifos_id", "loan_status", "term_frequency",
                                        "disbursed_on_date", "expected_matured_on_date",
                                        "closed_on_date"])

df[string_to_float_cols] = df[string_to_float_cols].apply(pd.to_numeric, errors='coerce')


#replace the in the column headers with underscores & make all column headers in lower case
df.columns = df.columns.str.lower().str.replace(' ','')

#create new column that concats the bloom version and loan id so as to prevent any loan duplicates from the same mifos instance

df["loan_id_product_concat"] = (df["loan_mifos_id"].astype("str")+"-"+df["bloom_version"].astype("str")).astype("str")

df.info()

#drop all loan records that have irrelevant status keys OR were never disbursed
df = df.drop(df[
    (df["loan_status"] == 0) |
    (df["loan_status"] == 100) |
    (df["loan_status"] == 400) |
    (df["loan_status"] == 500)
].index)


#drop all rows where loan id is blank
df = df.loc[df["loan_mifos_id"].notnull()]


#drop all rows with duplicate loan id is blank
df = df.loc[~df["loan_id_product_concat"].duplicated()]


#drop all rows where disbursement date is blank
df = df.loc[df["disbursed_on_date"].notnull()]

#clean up mobile number column
df[["client_mobile_number_2", "temp"]] = df["client_mobile_number"].astype("str").str.split(".", expand=True)


#drop newly created temp column
df.drop(["temp","client_mobile_number"], axis=1, inplace=True)

df.rename(columns={"client_mobile_number_2": "client_mobile_number"}, inplace=True)

#sort dataframe based on specific columns
df.sort_values(["client_mobile_number","disbursed_on_date"], ascending=[True, False], inplace=True)


#create loan count column i.e adds a new column that captures the num of loans a customer has taken
df["loan_count"] = df.groupby("store_number")["store_number"].transform('size')

#create loan rank column i.e adds a new column that captures the rank of each loan a customer has taken
df["loan_rank"] = df.groupby("store_number")["loan_mifos_id"].rank(ascending=True)

##sort out minor Mifos errors relating to loans with status 700
df.loc[(df["loan_status"] == 700) & (df["closed_on_date"].isnull()), "closed_on_date"] = df["expected_matured_on_date"]

## Notes ==> fix issue where loan rank gets messed up because of different Mifos instance

#create due_date fixed column to ensure due date tallies with the term frequency for each loan

df["due_date_fixed"] = df["disbursed_on_date"] + pd.to_timedelta(df["term_frequency"], unit='d')

#create a copy of df

all_loans = df.copy()

#analysis period

print('analysis begin date {}'.format(all_loans['disbursed_on_date'].min()))
print('analysis latest date {}'.format(all_loans['disbursed_on_date'].max()))

def calc_days_past_due(df):
    """
    Function to calculate days past due for each loan record. 
    The function uses pandas.Series vectorized arguments to ensure fast iterations/loops.
    Inputs are the arguments passed in the conditions list.
    Outputs are the results that are reported based on the choices list.
    Each output/choice assigned corresponds to the input/conditions level assigned above i.e.
    the first condition corresponds to the first choice etc, as such the rows of conditions & choices need to match.
    If the output reported is a string '0', that's an error/edgecase whose conditions, choices were not well declared.
    
    Inputs: 
    1) The current status of a loan as captured on corebanking(Mifos),
    2) The loans due date, 
    3) the last repayment date on record for the loan.
    
    
    Outputs:
    A calculation of a loan's number of days past due, that is converted from datetime to integer.
    
    """
    #subtract one day from today timestamp to ensure evaluation is in-line with Mifos which is Time-1 i.e one day behind
    today = (pd.Timestamp.today()).strftime('%Y-%m-%d')
    today = pd.Timestamp(today) - dt.timedelta(days=1)
        
    loan_status = df["loan_status"]
    due_date = df["due_date_fixed"]
    closed_on_date = df["closed_on_date"]
    
    #the main input is the loan status id that is used to slice the df
    conditions = [
        loan_status.eq(300),
        loan_status.eq(600),
        loan_status.eq(601),
        loan_status.eq(700),
    ]
    
    #the other date inputs are used for calculation based on the condition of loan status of a loan at any given point
    choices = [
        (today - due_date).dt.days,
        (closed_on_date - due_date).dt.days,
        (closed_on_date - due_date).dt.days,
        (closed_on_date - due_date).dt.days,
    ]
    
    days_past_due = np.select(conditions, choices)
    
    return days_past_due


#apply the function to the df to create the days_past_due_column
all_loans["days_past_due"] = calc_days_past_due(all_loans)

def set_loan_status_labels(df):
    """
    Function to set the loan repayment status of a loan.
    Inputs are the arguments passed in the conditions list.
    Outputs are the results that are reported based on the choices list.
    Each output/choice assigned corresponds to the input/conditions level assigned above i.e.
    the first condition corresponds to the first choice etc, as such the rows of conditions & choices need to match.
    If the output reported is a string '0', that's an error/edgecase whose conditions, choices were not well declared.
    
    Inputs: 
    1) The current status of a loan as captured on corebanking(Mifos),
    2) Number of days past due for each loan, 
    3) Term frequency for each loan,
    4) The version of Bloom tied to the loan record
    
    Outputs: 
    A string label noting the current loan repayment status of each loan record
    """
    thirty_day_product_rollover = 7 #Bloom 1.0
    twenty_one_day_product_rollover = 5 #Bloom 2.0 only
    seven_day_product_rollover_bloom1 = 7 #Bloom 2.0 adjusted to 5 days from 7 days in Bloom 1.0
    seven_day_product_rollover_bloom2 = 5 
    three_day_product_rollover = 2
    one_day_product_rollover = 1 #Bloom 2.0, may be adjusted to 5 days as well

    loan_status = all_loans["loan_status"]
    days_past_due = all_loans["days_past_due"]
    term_frequency = all_loans["term_frequency"]
    bloom_version = all_loans["bloom_version"]

    conditions = [
        #written off loans
        loan_status.eq(601),
        
        #currently active OR loans closed in tenure
        loan_status.eq(300) & days_past_due.le(0),
        loan_status.eq(600) & days_past_due.lt(0),
        loan_status.eq(700) & days_past_due.lt(0),
        loan_status.eq(600) & days_past_due.eq(0),
        loan_status.eq(700) & days_past_due.eq(0),
    
        #active loans that are presently in rollover
        loan_status.eq(300) & term_frequency.eq(1) & days_past_due.le(one_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(3) & days_past_due.le(three_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.le(seven_day_product_rollover_bloom1),
        loan_status.eq(300) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.le(seven_day_product_rollover_bloom2),
        loan_status.eq(300) & term_frequency.eq(21) & days_past_due.le(twenty_one_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(30) & days_past_due.le(thirty_day_product_rollover),
    
        #loans that were cleared/closed after they got to rollover and cleared with exact balance due
        loan_status.eq(600) & term_frequency.eq(1) & days_past_due.le(one_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(3) & days_past_due.le(three_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.le(seven_day_product_rollover_bloom1),
        loan_status.eq(600) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.le(seven_day_product_rollover_bloom2),
        loan_status.eq(600) & term_frequency.eq(21) & days_past_due.le(twenty_one_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(30) & days_past_due.le(thirty_day_product_rollover),
    
        #loans that were cleared/closed after they got to rollover and were overpaid
        loan_status.eq(700) & term_frequency.eq(1) & days_past_due.le(one_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(3) & days_past_due.le(three_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.le(seven_day_product_rollover_bloom1),
        loan_status.eq(700) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.le(seven_day_product_rollover_bloom2),
        loan_status.eq(700) & term_frequency.eq(21) & days_past_due.le(twenty_one_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(30) & days_past_due.le(thirty_day_product_rollover),

        #active loans that are presently in default
        loan_status.eq(300) & term_frequency.eq(1) & days_past_due.gt(one_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(3) & days_past_due.gt(three_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.gt(seven_day_product_rollover_bloom1),
        loan_status.eq(300) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.gt(seven_day_product_rollover_bloom2),
        loan_status.eq(300) & term_frequency.eq(21) & days_past_due.gt(twenty_one_day_product_rollover),
        loan_status.eq(300) & term_frequency.eq(30) & days_past_due.gt(thirty_day_product_rollover),


        #loans were cleared/closed when they had got to default status and cleared with exact balance due
        loan_status.eq(600) & term_frequency.eq(1) & days_past_due.gt(one_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(3) & days_past_due.gt(three_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.gt(seven_day_product_rollover_bloom1),
        loan_status.eq(600) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.gt(seven_day_product_rollover_bloom2),
        loan_status.eq(600) & term_frequency.eq(21) & days_past_due.gt(twenty_one_day_product_rollover),
        loan_status.eq(600) & term_frequency.eq(30) & days_past_due.gt(thirty_day_product_rollover),

        #loans that were cleared/closed after they got to default and were overpaid
        loan_status.eq(700) & term_frequency.eq(1) & days_past_due.gt(one_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(3) & days_past_due.gt(three_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(7) & bloom_version.eq(1) & days_past_due.gt(seven_day_product_rollover_bloom1),
        loan_status.eq(700) & term_frequency.eq(7) & bloom_version.eq(2) & days_past_due.gt(seven_day_product_rollover_bloom2),
        loan_status.eq(700) & term_frequency.eq(21) & days_past_due.gt(twenty_one_day_product_rollover),
        loan_status.eq(700) & term_frequency.eq(30) & days_past_due.gt(thirty_day_product_rollover),
    ]


    choices = [
        "written-off_default",
        
        "current_active",
        "closed_early_repayment",
        "closed_early_repayment_overpaid",
        "closed_on_time",
        "closed_on_time_overpaid",    
        
        "active_rollover",
        "active_rollover",
        "active_rollover",
        "active_rollover",
        "active_rollover",
        "active_rollover",
        
        "closed_rollover",
        "closed_rollover",
        "closed_rollover",
        "closed_rollover",
        "closed_rollover",
        "closed_rollover",    
        
        "closed_rollover_overpaid",
        "closed_rollover_overpaid",
        "closed_rollover_overpaid",
        "closed_rollover_overpaid",
        "closed_rollover_overpaid",
        "closed_rollover_overpaid",
        
        "active_default",
        "active_default",
        "active_default",
        "active_default",
        "active_default",
        "active_default",   
        
        "closed_default",
        "closed_default",
        "closed_default",
        "closed_default",
        "closed_default",
        "closed_default",        
        
        "closed_default_overpaid",
        "closed_default_overpaid",
        "closed_default_overpaid",
        "closed_default_overpaid",
        "closed_default_overpaid",
        "closed_default_overpaid",

    ]
    
    loan_labels = np.select(conditions, choices)
    
    return loan_labels


#apply the function to the df to create the days_past_due_column
all_loans["loan_repayment_status"] = set_loan_status_labels(all_loans)

#return df with two most recent loans for each client id

#temp_df = all_loans.groupby("client_id").head(2).reset_index()

all_loans.head(2)

all_loans.shape

all_loans['store_number'].nunique()

#convert col to string
all_loans["store_number"] = all_loans["store_number"].astype("str")


#clean up store number column
all_loans[["store_number", "temp"]] = all_loans["store_number"].astype("str").str.split(".", expand=True)


#drop newly created temp column
all_loans.drop(["temp",], axis=1, inplace=True)

all_loans.head(2)

all_loans.info()

#all_loans= all_loans[all_loans['disbursed_on_date'] <= '2022-06-14']

---
#### Generate aggregate summaries

#return df with the most recent loan for each borrower

agg_summary = all_loans.groupby("store_number").head(1).reset_index()

agg_summary.shape

agg_summary['store_number'].nunique()

#trim df to only relevant columns
target_cols = ["client_mobile_number","store_number","loan_count","loan_status","term_frequency",
  "principal_disbursed","principal_repaid","disbursed_on_date",
  "expected_matured_on_date","closed_on_date","due_date_fixed",
  "days_past_due","bloom_version","loan_repayment_status"]

agg_summary = agg_summary[target_cols]

#aggregate maximum principal disbursed for each client id

agg_summary = pd.merge(agg_summary, (all_loans.groupby("store_number")["principal_disbursed"].max().rename("max_principal_amount").reset_index()), on="store_number")

#get df for when a customer got their max loan principal

max_principal_dates = all_loans.sort_values("principal_disbursed", ascending=False).groupby("store_number").first().reset_index()

#trim df to only remain with relevant columns

max_principal_dates = max_principal_dates[["store_number","disbursed_on_date"]]

#rename column to make it clearer
max_principal_dates.rename(columns = {"disbursed_on_date": "max_loan_disbursement_date"}, inplace = True)

#merge df

agg_summary = pd.merge(agg_summary, max_principal_dates, on="store_number")

#aggregate of loans well paid or in good standing

agg_good_loans = (all_loans.loc[(all_loans["loan_repayment_status"]== "closed_early_repayment")|
                                (all_loans["loan_repayment_status"]== "closed_early_repayment_overpaid")|
                                (all_loans["loan_repayment_status"]== "closed_on_time")|
                                (all_loans["loan_repayment_status"]== "closed_on_time_overpaid")|
                                (all_loans["loan_repayment_status"]== "current_active")
                               ]).groupby("store_number")["loan_id_product_concat"].aggregate("count").rename("count_good_loans").reset_index()

#merge agg_summary & agg good loans

agg_summary = pd.merge(agg_summary, agg_good_loans, how="outer", on="store_number")

#fill rest of missing values with zeros

cols_fillna = ['count_good_loans']
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    agg_summary[col].fillna(0,inplace=True)

#calculate good repayment ratios for the borrowers

agg_summary["good_loans_repayment_ratio"] = round(agg_summary["count_good_loans"]/agg_summary["loan_count"], 2)

#calculate num days since last disbursement
#subtract one day from today timestamp to ensure evaluation is in-line with Mifos which is Time-1 i.e one day behind

today = (pd.Timestamp.today()).strftime('%Y-%m-%d')
today = pd.Timestamp(today) - dt.timedelta(days=1)

agg_summary["num_days_since_last_disbursement"] = pd.to_numeric((today - agg_summary["disbursed_on_date"]).dt.days, downcast='integer')

#delete test accounts

agg_summary = agg_summary[agg_summary["max_principal_amount"]>=200]

#delete customers without a store number

agg_summary = agg_summary.loc[agg_summary["store_number"].notnull()]

#add column for max dpd

agg_summary = pd.merge(agg_summary, (all_loans.groupby("store_number")["days_past_due"].max().rename("max_days_past_due").reset_index()), on="store_number")

agg_summary.head(2)

agg_summary.shape

---
#### slice of DF with customers who have a max 30DPD

#get a slice of df where customers have a max dpd of 30 days past due

max_dpd_30 = agg_summary.loc[agg_summary["max_days_past_due"]<=30]

max_dpd_30.head(3)

max_dpd_30_list = list(max_dpd_30["store_number"].unique())

len(max_dpd_30_list)

---
#### get slice of df for loans over the past 3 months i.e 180 days

#filter df for 6+ loans & past 90 days

current_period = all_loans["disbursed_on_date"].max()
target_analysis_period = current_period - pd.DateOffset(months=3)
min_loan_amount = 200

loans_past_3_months = all_loans.loc[(all_loans["disbursed_on_date"]>=target_analysis_period)&
                                    (all_loans["principal_disbursed"]>=min_loan_amount)
                                   ]

all_loans.head(3)

all_loans.shape

#drop rows where store number is null

loans_past_3_months = loans_past_3_months.loc[loans_past_3_months["store_number"].notnull()]

#calculate num of loans taken past 3 months

loans_past_3_months["loan_count_past_3_months"] = loans_past_3_months.groupby("store_number")["store_number"].transform('size')

#create loan count column i.e adds a new column that captures the num of 7-day loans a customer has taken
loans_past_3_months["count_7_day_loans"] = loans_past_3_months.loc[loans_past_3_months["term_frequency"]==7].groupby("store_number")["store_number"].transform('size')

#fill rest of missing values with zeros

cols_fillna = ["count_7_day_loans"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months[col].fillna(0,inplace=True)

#how many customers have taken a loan over the past

loans_past_3_months["store_number"].nunique()

loans_past_3_months.head(2)

loans_past_3_months_stores = loans_past_3_months.groupby(['store_number'], as_index=False)['loan_count_past_3_months'].max()
loans_past_3_months_stores.head(3)

loans_past_3_months_stores.shape

#get list of all store numbers

loans_past_3_months_list = list(loans_past_3_months["store_number"].unique())

---
##### get agg summaries for past 3 months 7-day loans summaries

#get aggregate summary with avg loan tenure

loans_past_3_months_summary = round(loans_past_3_months.groupby("store_number")["term_frequency"].aggregate("mean").rename("avg_loan_tenure").reset_index(), 0)

#get summary of number of 7 day loans a customer has taken

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, loans_past_3_months.groupby("store_number")["count_7_day_loans"].max().rename("count_7_day_loans").reset_index(), how="left", on="store_number")

#merge df to get number of 7 day loans paid upto rollover

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, (loans_past_3_months.loc[loans_past_3_months["term_frequency"]==7].loc[
    (loans_past_3_months["loan_repayment_status"]== "closed_early_repayment")|
    (loans_past_3_months["loan_repayment_status"]== "closed_early_repayment_overpaid")|
    (loans_past_3_months["loan_repayment_status"]== "closed_on_time")|
    (loans_past_3_months["loan_repayment_status"]== "closed_on_time_overpaid")|
    (loans_past_3_months["loan_repayment_status"]== "current_active")|
    (loans_past_3_months["loan_repayment_status"]== "active_rollover")|
    (loans_past_3_months["loan_repayment_status"]== "closed_rollover")|
    (loans_past_3_months["loan_repayment_status"]== "closed_rollover_overpaid")
]).groupby("store_number")["client_mifos_id"].aggregate("count").rename("count_7_day_loans_paid_upto_rollover").reset_index(), how="left", on="store_number")

#fill rest of missing values with zeros

cols_fillna = ["count_7_day_loans_paid_upto_rollover"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

#calculate good repayment ratios for the borrowers

loans_past_3_months_summary["good_loans_repayment_ratio(7_day_loans)"] = round(loans_past_3_months_summary["count_7_day_loans_paid_upto_rollover"]/loans_past_3_months_summary["count_7_day_loans"], 2)

#fill rest of missing values with zeros

cols_fillna = ["good_loans_repayment_ratio(7_day_loans)"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

# get minimum amount disbursed for a 7 day loan

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, loans_past_3_months.loc[loans_past_3_months["term_frequency"]==7].groupby("store_number")["principal_disbursed"].min().rename("minimum_7_day_principal_disbursed").reset_index(), how="left", on="store_number")

# get total amount disbursed for a 7 day loan

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, loans_past_3_months.loc[loans_past_3_months["term_frequency"]==7].groupby("store_number")["principal_disbursed"].sum().rename("total_sum_7_day_principal_disbursed").reset_index(), how="left", on="store_number")

# get total amount disbursed for a 1 day loan

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, loans_past_3_months.loc[loans_past_3_months["term_frequency"]==1].groupby("store_number")["principal_disbursed"].sum().rename("total_sum_1_day_principal_disbursed").reset_index(), how="left", on="store_number")

#fill rest of missing values with zeros

cols_fillna = ["minimum_7_day_principal_disbursed"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

#fill rest of missing values with zeros

cols_fillna = ["total_sum_7_day_principal_disbursed"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

#fill rest of missing values with zeros

cols_fillna = ["total_sum_1_day_principal_disbursed"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

#get the average 7 day loan principal ==> proxy for 7 day limit

loans_past_3_months_summary = pd.merge(loans_past_3_months_summary, loans_past_3_months.loc[loans_past_3_months["term_frequency"]==7].groupby("store_number")["principal_disbursed"].mean().rename("avg_7_day_principal_disbursed").reset_index(), how="left", on="store_number")

#fill rest of missing values with zeros

cols_fillna = ["avg_7_day_principal_disbursed"]
# replace 'NaN' with zero in these columns
for col in cols_fillna:
    loans_past_3_months_summary[col].fillna(0,inplace=True)

def approximate_21_day_limit(df):
    """
    """
    
    seven_day_limit = df["avg_7_day_principal_disbursed"]
    seven_day_limit_factor = 0.17
    twenty_one_day_limit_factor = 0.5
    product_cap = 200000
    zero = 0
    
    operation = (np.ceil(((df["avg_7_day_principal_disbursed"]*twenty_one_day_limit_factor)/seven_day_limit_factor)/100) * 100).astype(int)
    
    conditions = [
        seven_day_limit.eq(zero),
        operation.le(product_cap),
        operation.gt(product_cap),
    ]
    
    choices = [
        zero,
        operation,
        product_cap,
    ]
    
    limit_col = np.select(conditions, choices)
    
    return limit_col

#execute function

loans_past_3_months_summary["21_day_limit"] = approximate_21_day_limit(loans_past_3_months_summary)

loans_past_3_months_summary.shape

loans_past_3_months_summary.head()

#### num of customers with loans repayment ratio of >=90% for their 7 day loans over the past 3 months

loans_past_3_months_summary.loc[loans_past_3_months_summary["good_loans_repayment_ratio(7_day_loans)"]>=0.9]["store_number"].nunique()

# num of customers with total 1 day sum of principal > 35K

loans_past_3_months_summary.loc[loans_past_3_months_summary["total_sum_7_day_principal_disbursed"]>=35000]["store_number"].nunique()

loans_past_3_months_summary.loc[loans_past_3_months_summary["count_7_day_loans"]>=3]["store_number"].nunique()

# num of customers with 21 day exceeding 7500

loans_past_3_months_summary.loc[loans_past_3_months_summary["21_day_limit"]>=7500]["store_number"].nunique()

# num of customers with an average loan tenure greater than 1

loans_past_3_months_summary.loc[loans_past_3_months_summary["avg_loan_tenure"]>1]["store_number"].nunique()

loans_past_3_months_summary.head()

loans_past_3_months_summary.head(2)

loans_past_3_months_summary['total_sum_7_day_principal_disbursed'].max()

loans_past_3_months_summary.info()

---
#### Fetch latest limits data from warehouse

# Connect to the database and load the scoring data bloomlive table
conn = connect(param_dic)

columns = "store_number,final_21_limit, final_7_limit, final_1_limit"

column_names = columns.strip().split(",")
# Execute the "SELECT cols" query
df = postgresql_to_dataframe(conn,
                             "SELECT \
                             store_number, final_21_limit, final_7_limit, final_1_limit  \
                             FROM bloomlive.scoring_results where model_version ='2022-004[2022-05-14, 2022-06-29]'",
                             column_names)

df = pd.read_excel("/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220922_multiple_products_2.xlsx")

df = df[['store_number', 'final_21_limit', 'final_7_limit', 'final_1_limit']]

df.shape

df[['final_21_limit', 'final_7_limit', 'final_1_limit']].sum()

#replace the in the column headers with underscores & make all column headers in lower case
df.columns = df.columns.str.lower().str.replace(' ','')

df.info()

#convert specific string columns to float, first by leaving out the non-targeted columns
string_to_float_cols = df.columns.drop(["final_21_limit", "final_7_limit", "final_1_limit"])

df[string_to_float_cols] = df[string_to_float_cols].apply(pd.to_numeric, errors='coerce')

df.shape

## get slice of df where customer has 7 day limit but 21 day limit is 0

#slice_df = df.loc[(df["final_7_limit"]>200)&(df["final_21_limit"]==0)]

slice_df = df.loc[df["final_21_limit"]>0]

slice_df.shape

#convert store num col to str

slice_df["store_number"] = slice_df["store_number"].astype("str")

## num customers that have 7 day limit but 21 day limit is 0

slice_df["store_number"].nunique()

#get list of all target store numbers

slice_df_list = list(slice_df["store_number"].unique())

len(slice_df_list)

#df["store_number"].nunique()

df_list = list(df["store_number"].unique())

len(df_list)

---
#### Applying all factors to get the final df

#filter loans summary df to see how many customers had 7 day limit but no 21 day limit

target_customers_scope = loans_past_3_months_summary.loc[loans_past_3_months_summary["store_number"].isin(slice_df_list)]

print(target_customers_scope.shape)

target_customers_scope.head(3)

#check to make sure customer has no history of >30 DPD

ntarget_customers_scope = target_customers_scope.loc[target_customers_scope["store_number"].isin(max_dpd_30_list)]

print(ntarget_customers_scope.shape)

ntarget_customers_scope.loc[(ntarget_customers_scope["good_loans_repayment_ratio(7_day_loans)"]>=0.9)].shape

ntarget_customers_scope.loc[(ntarget_customers_scope["total_sum_7_day_principal_disbursed"]>=35000)].shape

ntarget_customers_scope.loc[(ntarget_customers_scope["count_7_day_loans"]>=3)].shape

ntarget_customers_scope.loc[(ntarget_customers_scope["21_day_limit"]>=7500)].shape

#all filters applied

mtarget_customers_scope = ntarget_customers_scope.loc[
    (ntarget_customers_scope["good_loans_repayment_ratio(7_day_loans)"]>=0.9)& 
    (ntarget_customers_scope["total_sum_7_day_principal_disbursed"]>=25000)&
    (ntarget_customers_scope["count_7_day_loans"]>=3)&
    (ntarget_customers_scope["21_day_limit"]>=7500)
]

print(mtarget_customers_scope.shape)

#all filters applied

mtarget_customers_scope = ntarget_customers_scope.loc[
    (ntarget_customers_scope["good_loans_repayment_ratio(7_day_loans)"]>=0.9)&
    (ntarget_customers_scope["21_day_limit"]>=7500)
]

print(mtarget_customers_scope.shape)

mtarget_customers_scope.shape

mtarget_customers_scope.head(3)

qualified_stores = mtarget_customers_scope[['store_number']]

df['store_number'] = df['store_number'].astype(str)

final = pd.merge(qualified_stores, df, on = 'store_number', how = 'left')

final.head()

final.shape

final['final_21_limit'].sum()

final_21_day = final[['store_number', 'final_21_limit']]

final_21_day.head()

df_adj = df[['store_number', 'final_7_limit', 'final_1_limit']]

df_adj.head()

final_limits = pd.merge(final_21_day, df_adj, on = 'store_number', how = 'outer')

final_limits.head()

final_limits.shape

final_limits[['final_21_limit', 'final_7_limit', 'final_1_limit']].sum()

final_limits.to_excel('final_limits_new.xlsx')
