==> notebooks/17.1_JN_zeroizing_7_day_and_1_day_limits_for_merchants_with_saf_score_0-477.ipynb <==
import pandas as pd
import numpy as np

import psycopg2
import pymysql

from datetime import timedelta

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

pd.set_option('display.float_format', lambda x: '%.3f' % x)

df = pd.read_excel("/home/jupyter-ngenia-jackline/comb_limits_new.xlsx")
#df = pd.read_excel("/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220908_multiple_products_rec.xlsx")
df = df.rename(columns={'final_21_limits':'new_final_21_limit', 'bf_final_7_limit':'new_final_7_limit', 'bf_final_1_limit':'new_final_1_limit'})

df.head()

df[['new_final_21_limit', 'new_final_7_limit', 'new_final_1_limit']].sum()

df.info()

below_score = df[(df['old_src_crdt_score'] >= 0) & (df['old_src_crdt_score'] <= 477)]

below_score[['new_final_21_limit', 'new_final_7_limit', 'new_final_1_limit']].sum()

below_score.to_excel('below_score.xlsx')

def calc_final_7_limit(df):
    
    old_src_crdt_score = df['old_src_crdt_score']
    new_final_7_limit = df['new_final_7_limit']
    
    if old_src_crdt_score >= 0 and old_src_crdt_score <=477:
        return 0
    else:
        return new_final_7_limit
    

    
def calc_final_1_limit(df):
    
    old_src_crdt_score = df['old_src_crdt_score']
    new_final_1_limit = df['new_final_1_limit']
    
    if old_src_crdt_score >= 0 and old_src_crdt_score <=477:
        return 0
    else:
        return new_final_1_limit

#df['final_21_limit'] = df.apply(lambda x: calc_final_21_limit(x), axis = 1)
df['updated_final_7_limit'] = df.apply(lambda x: calc_final_7_limit(x), axis = 1)
df['updated_final_1_limit'] = df.apply(lambda x: calc_final_1_limit(x), axis = 1)

df[['new_final_21_limit', 'updated_final_7_limit', 'updated_final_1_limit']].sum()

df[(df['old_src_crdt_score'] >= 0) & (df['old_src_crdt_score'] <= 477)].shape

df['old_src_crdt_score'].describe()

df.head()

df.to_excel('/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220922_multiple_products_rec_new.xlsx')

df.shape
