==> notebooks/19.1_JN_adding_back_7_day_and_1_day_limits_for_merchants_with_only_21_day_limits.ipynb <==
import pandas as pd
import numpy as np
import datetime

import psycopg2
import pymysql

from datetime import timedelta

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

pd.set_option('display.float_format', lambda x: '%.3f' % x)

#current_refresh = pd.read_excel("/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220901_multiple_products_rec.xlsx")
current_refresh = pd.read_excel("/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220922_multiple_products_rec_1_new.xlsx")
#current_refresh = current_refresh[['store_number', 'new_final_21_limit', 'updated_7_limit', 'updated_1_limit']]
#current_refresh = current_refresh.rename(columns={'updated_7_limit':'new_final_7_limit','updated_1_limit':'new_final_1_limit'})

#current_refresh = current_refresh.fillna(0)

current_refresh.head()

current_refresh[(current_refresh['new_final_21_limit_1'] > 0) & (current_refresh['updated_7_limit'] == 0)].shape

def updated_new_7_limit(df):
    
    updated_7_limit = df['updated_7_limit']
    new_final_21_limit_1 = df['new_final_21_limit_1']
    old_final_7_limit = df['old_final_7_limit']
    
    if new_final_21_limit_1 > 0 and updated_7_limit == 0:
        return old_final_7_limit
    else:
        return updated_7_limit
    
    
def updated_new_1_limit(df):
    
    updated_1_limit = df['updated_1_limit']
    new_final_21_limit_1 = df['new_final_21_limit_1']
    old_final_1_limit = df['old_final_1_limit']
    
    if new_final_21_limit_1 > 0 and updated_1_limit == 0:
        return old_final_1_limit
    else:
        return updated_1_limit

current_refresh['updated_new_21_limit'] = current_refresh['new_final_21_limit_1']
current_refresh['updated_new_7_limit'] = current_refresh.apply(lambda x: updated_new_7_limit(x), axis = 1)
current_refresh['updated_new_1_limit'] = current_refresh.apply(lambda x: updated_new_1_limit(x), axis = 1)

print(current_refresh[['updated_new_21_limit', 'updated_new_7_limit', 'updated_new_1_limit']].sum())

current_refresh.to_excel("/home/jupyter-ngenia-jackline/Limits_refresh_summary_20220922_multiple_products_rec_2_new.xlsx")
