==> notebooks/20.4_FM_bloom_first_time_defaults_monitoring.ipynb <==
# Import modules
import sys

# Import custom modules
sys.path.append('../')

from src.models.upload_data import *

# Set notebook attributes
pd.set_option('display.max_colwidth', 500)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 300)
pd.set_option('display.float_format', lambda x: '%.2f' % x)

# Set magic attributes
%matplotlib inline
%load_ext autoreload
%autoreload 2

# Configuration
args = argparse.ArgumentParser()
args.add_argument("--config", default="../params.yaml")
parsed_args = vars(args.parse_args(""))
config_path = parsed_args['config']

config = read_params(config_path)
project_dir = config["project_dir"]
ftd_clean_data_path = config["processed_data_config"]["ftd_clean_data_parquet"]
scored_limits_risk_review_data_path_parquet = config["processed_data_config"]["scored_limits_risk_review_data_parquet"]

### First Time Defaulters Monitoring

#### Latest Limits

%%time
# Data loading
df_scoring_stabilisation = pd.read_parquet(project_dir + scored_limits_risk_review_data_path_parquet)
df_scoring_stabilisation.sample(2)

# Data set dimension
df_scoring_stabilisation.shape

#### First Time Defaults

%%time
# Data loading
df_ftd_clean = pd.read_parquet(project_dir + ftd_clean_data_path)
df_ftd_clean.sample(2)

# Data set dimension
df_ftd_clean.shape

# Check unique values
df_ftd_clean.store_number.nunique()

#### Merge Datasets

# Columns in-scope
cols_in_scope = ['store_number', 'final_21_limit', 'final_7_limit', 'final_1_limit', 
                 "model_630_21_limit", "model_630_7_limit", "model_630_1_limit", "model_630_max_global_limit", 
                 "loan_balance", "blacklist_flag"]

# Merge datasets
ftd_monitoring = df_scoring_stabilisation.merge(df_ftd_clean['store_number'], how="right", on=["store_number"])
ftd_monitoring.sample(2)

# Data set dimension
ftd_monitoring.shape

#### Diagnostics

# With limits
wl = ftd_monitoring.loc[(ftd_monitoring['blacklist_flag'] == 0)]
print(wl.shape)
wl[cols_in_scope].sample(2)

# Without limits not in model_630
wol_ms = ftd_monitoring.loc[(ftd_monitoring['blacklist_flag'] == 1) & (ftd_monitoring['model_630_21_limit'].isna())]
print(wol_ms.shape)
wol_ms[cols_in_scope].head(2)

# Without limits balance greater than limit
wol_l = ftd_monitoring.loc[(ftd_monitoring['blacklist_flag'] == 1) & ~(ftd_monitoring['model_630_21_limit'].isna())]
print(wol_l.shape)
wol_l[cols_in_scope].sample(2)

# Without limits
inv = ftd_monitoring.loc[(ftd_monitoring['blacklist_flag'].isna())]
print(inv.shape)
inv[cols_in_scope].head()

# Data set dimension
wl.shape[0] + wol_ms.shape[0] + wol_l.shape[0] + inv.shape[0]
