==> notebooks/53.2_FM_bloom_limit_investigation.ipynb <==
# Import modules
import sys

# Import custom modules
sys.path.append('../')

from src.models.upload_data import *

# Set notebook attributes
pd.set_option('display.max_colwidth', 500)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', 300)
pd.set_option('display.float_format', lambda x: '%.2f' % x)

# Set magic attributes
%matplotlib inline
%load_ext autoreload
%autoreload 2

# Configuration
args = argparse.ArgumentParser()
args.add_argument("--config", default="../params.yaml")
parsed_args = vars(args.parse_args(""))
config_path = parsed_args['config']

# Load configurations
config = read_params(config_path)
project_dir = config["project_dir"]
scored_limits_risk_review_data_path_excel_parquet = config["processed_data_config"]["scored_limits_risk_review_data_excel_parquet"]
before_21d_graduation_limits_data_path = config["interim_data_config"]["before_21d_graduation_limits_data_parquet"]
after_rmdd_ge95_data_path = config["interim_data_config"]["after_rmdd_ge95_data_parquet"]
after_rmdd_ge96_data_path = config["interim_data_config"]["after_rmdd_ge96_data_parquet"]

### Limit Investigation

#### Legacy

# Columns not in scope
df_columns_not_in_scope = ['sum_3_months_trx_val', 'approx_30_days_trx_val', 'model_version', 'created_at', 'record_added_to_warehouse_on_timestamp', 'old_src_crdt_score']
limit_cols = ['final_21_limit', 'final_7_limit', 'final_1_limit']

# Legacy system refresh run
# df = pd.read_excel('../data/external/07_JN_Limits_refresh_summary_20221006_multiple_products_rec_new_2.xlsx')
# df.to_parquet('../data/external/07_JN_Limits_refresh_summary_20221006_multiple_products_rec_new_2.parquet', index=False)
df_raw = pd.read_parquet('../data/external/07_JN_Limits_refresh_summary_20221006_multiple_products_rec_new_2.parquet')
df_raw['store_number'] = df_raw['store_number'].astype('str')
df = df_raw.drop(columns=df_columns_not_in_scope)
df.rename(columns={'avg_3_months_trx_val': 'approx_30_days_trx_val'}, inplace=True)
df.sample(2)

# Data set dimension
df.shape

# Data set schema
df.dtypes.to_frame().T

#### Ubuntu 2.0

# Ubuntu 2.0 refresh run
df_scoring_stabilisation_raw = pd.read_parquet(project_dir + scored_limits_risk_review_data_path_excel_parquet.format(refresh_date.replace("-", "")))
df_scoring_stabilisation = df_scoring_stabilisation_raw[df.columns]
df_scoring_stabilisation.sample(2)

# Data set dimension
df_scoring_stabilisation.shape

# Data set schema
df_scoring_stabilisation.dtypes.to_frame().T

#### Comparison

# In legacy and not in Ubuntu 2.0
df_diff_in_legacy = df[~(df['store_number'].isin(df_scoring_stabilisation['store_number']))]
df_diff_in_legacy.shape

# In Ubuntu 2.0 and not in legacy
df_diff_in_ubuntu = df_scoring_stabilisation[~(df_scoring_stabilisation['store_number'].isin(df['store_number']))]
df_diff_in_ubuntu.shape

# In legacy and in Ubuntu 2.0
df_similar = df[(df['store_number'].isin(df_scoring_stabilisation['store_number']))]
df_similar.shape

#### Samples

# Sample similar in legacy
df_similar_l = df[df['store_number'].isin(df_similar['store_number'])]
df_similar_ls = df_similar_l.sample(2).sort_values(by=['store_number'])
df_similar_l[df_similar_l['store_number'].isin(['7361329', '943478', '7252207'])].sort_values(by=['store_number'])

# Sample similar in Ubuntu 2.0
df_similar_u = df_scoring_stabilisation[df_scoring_stabilisation['store_number'].isin(df_similar_l['store_number'])]
df_similar_us = df_similar_u[df_similar_u['store_number'].isin(df_similar_ls['store_number'])]
df_similar_u[df_similar_u['store_number'].isin(['7361329', '943478', '7252207'])].sort_values(by=['store_number'])

#### Limits

# Legacy limits
df[limit_cols].sum()

# Ubuntu 2.0 limits
df_scoring_stabilisation[limit_cols].sum()

#### Diagnostics

##### 01. approx_30_days_trx_val

This feature is understated by 1 day in legacy. 

##### 02. most_recent_trx_date_past_30_days

# Legacy
df.last_trx_date.max() - df.last_trx_date.min()

# Ubuntu 2.0
df_scoring_stabilisation.last_trx_date.max() - df_scoring_stabilisation.last_trx_date.min()

Legacy looks at a 29 day window.
Ubuntu 2.0 looks at a 30 day window.

This affects:

* expected_trx_days = last_trx_date - most_recent_trx_date_past_30_days `Shorter window in scope`
* actual_trx_days = COUNT UNIQUE transaction_time `Shorter window in scope`
* page_active_days = actual_trx_days / expected_trx_days `Computational metrics are under respresented`
* inference_col

##### 03. days_since_last_trx

df.loc[(df.last_trx_date == (extract_end_date - dt.timedelta(days=5))), 'days_since_last_trx'].unique()

df_scoring_stabilisation[df_scoring_stabilisation.days_since_last_trx == 5].head()

This feature is ovestated by 1 day in legacy. 
Refresh date is factored in computation logic.

days_since_last_trx = extract_end_date - last_trx_date

This affects:

* transacted_last_5_days `Actual window not in scope`

##### 04. good_loans_repayment_ratio

Some values are null in Ubuntu 2.0.
This is beacuse of a loan count of 0.

This has been remediated by imputing missing values.

This affects:

* inference_col
* transacted_last_5_days
* weight_till_recency

##### 05. max_loan_disbursement_date

Ought to be sorted by both amount and disbured_on_date due to duplicate max principal amount.

##### 06. loan_id_product_concat

Bloom version for td data set ought to be coersed to float

##### 07. weight_recency

Threshold of 3
Less than or equal to three return 1
Else 0

##### 08. old_src_credit_score

Definition not available.
