==> notebooks/56.1_JN_Bloom_Regression_Model.ipynb <==
# **Impact Assessment**

### Import processing and prepping libraries

#import data processing libraries
import pandas as pd
import numpy as np
import psycopg2
import pymysql
import datetime as dt
from datetime import timedelta
import os
import math as math
from scipy import stats

#import visualization libraries
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
%matplotlib inline

from sklearn.cluster import KMeans, AgglomerativeClustering
from sklearn.linear_model import LinearRegression
from sklearn.impute import SimpleImputer
from sklearn.tree import DecisionTreeRegressor
from sklearn.preprocessing import StandardScaler, normalize, LabelEncoder
from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from sklearn.metrics import silhouette_score, classification_report, confusion_matrix, mean_squared_error, r2_score, mean_absolute_error
from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
from sklearn.manifold import TSNE
import xgboost as xgb
from sklearn.svm import SVR

import joblib
import pyarrow as pa
import pickle
from umap import UMAP

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

pd.set_option('display.float_format', lambda x: '%.3f' % x)

### Load the Data

# repayments = pd.read_csv("repayments_by_different_milestones.csv")
# repayments.to_parquet('repayments_by_different_milestones_2.parquet', index=False)

repayments = pd.read_parquet('repayments_by_different_milestones_2.parquet')

repayments['store_number'] = repayments['store_number'].astype(str)

repayments.head(3)

repayments = repayments[(repayments['term_frequency'] == 7) | (repayments['term_frequency'] == 21)]

repayments.info()

repayments['repayment_amount_by_due_date'] = repayments['repayment_amount_by_due_date'].fillna(0)
repayments['repayment_amount_by_rllvr_date'] = repayments['repayment_amount_by_rllvr_date'].fillna(0)
repayments['repayment_amount_by_dpd30'] = repayments['repayment_amount_by_dpd30'].fillna(0)

repayments.info()

repayments['repayment_rate_by_due_date'] = round(repayments['repayment_amount_by_due_date'] / repayments['principal_disbursed'], 4)
repayments['repayment_rate_by_rllvr_date'] = round(repayments['repayment_amount_by_rllvr_date'] / repayments['principal_disbursed'], 4)
repayments['repayment_rate_by_dpd30'] = round(repayments['repayment_amount_by_dpd30'] / repayments['principal_disbursed'], 4)

repayments.head(3)

hurdle_rate_by_due_date = repayments.groupby(['store_number'], as_index=False)['repayment_rate_by_due_date'].mean()
hurdle_rate_by_due_date = hurdle_rate_by_due_date.rename(columns={'repayment_rate_by_due_date':'hurdle_rate_by_due_date_mean'})

hurdle_rate_by_rllvr_date = repayments.groupby(['store_number'], as_index=False)['repayment_rate_by_rllvr_date'].mean()
hurdle_rate_by_rllvr_date = hurdle_rate_by_rllvr_date.rename(columns={'repayment_rate_by_rllvr_date':'hurdle_rate_by_end_rollover_date_mean'})

hurdle_rate_by_dpd30 = repayments.groupby(['store_number'], as_index=False)['repayment_rate_by_dpd30'].mean()
hurdle_rate_by_dpd30 = hurdle_rate_by_dpd30.rename(columns={'repayment_rate_by_dpd30':'hurdle_rate_by_dpd30_mean'})

hurdle_rates_df = pd.merge(hurdle_rate_by_due_date, hurdle_rate_by_rllvr_date, on='store_number', how='outer')
hurdle_rates_df = pd.merge(hurdle_rates_df, hurdle_rate_by_dpd30, on='store_number', how='outer')

hurdle_rates_df = hurdle_rates_df.fillna(0)

hurdle_rates_df.head(3)

# current_refresh = pd.read_excel("Limits_refresh_summary_20230817_multiple_products.xlsx")
# current_refresh.drop(columns=['Unnamed: 0'], inplace=True)
# current_refresh.to_parquet("Limits_refresh_summary_20230817_multiple_products.parquet", index=False)

current_refresh = pd.read_parquet("Limits_refresh_summary_20230914_multiple_products.parquet")

del current_refresh['hurdle_rate_by_due_date_mean']
del current_refresh['hurdle_rate_by_end_rollover_date_mean']
del current_refresh['hurdle_rate_by_dpd30_mean']

current_refresh['store_number'] = current_refresh['store_number'].astype(str)

current_refresh.head(2)

current_refresh = pd.merge(current_refresh, hurdle_rates_df, on='store_number', how='left')

current_refresh.head(2)

##############################
current_refresh['days_since_last_trx'] = current_refresh['days_since_last_trx'].fillna(31)

current_refresh[['final_21_limit', 'final_7_limit', 'final_1_limit']].sum()

# selected_columns = current_refresh[["store_number", "approx_30_days_trx_val", "expected_trx_days", "actual_trx_days", "page_active_days", "days_since_last_trx", "weight_consistency", "weight_recency", "hurdle_rate_by_due_date_mean", "hurdle_rate_by_end_rollover_date_mean", "hurdle_rate_by_dpd30_mean"]]
selected_columns = current_refresh[["store_number", "approx_30_days_trx_val", "expected_trx_days", "actual_trx_days", "page_active_days", "days_since_last_trx", "weight_consistency", "weight_recency", "hurdle_rate_by_dpd30_mean"]]

selected_columns.head(2)

selected_columns.shape

without_performance_data = selected_columns[selected_columns['hurdle_rate_by_dpd30_mean'].isnull()].reset_index(drop=True)
print(without_performance_data.shape)
with_performance_data = selected_columns[selected_columns['hurdle_rate_by_dpd30_mean'].notnull()].reset_index(drop=True)
print(with_performance_data.shape)

with_performance_data[with_performance_data['hurdle_rate_by_dpd30_mean'] >= 1.03].shape

(11704/32082) * 100

summaries = with_performance_data.copy()

# Assuming 'X' contains your independent variables (excluding the unique identifier) and 'y' contains the target variable
X_with_performance_data = with_performance_data.drop(['hurdle_rate_by_dpd30_mean', 'store_number'], axis=1)
y_with_performance_data = with_performance_data['hurdle_rate_by_dpd30_mean']

# Split the dataset into a training set and a test set (adjust the test_size as needed)
X_train_with_performance_data, X_test_with_performance_data, y_train_with_performance_data, y_test_with_performance_data = train_test_split(X_with_performance_data, y_with_performance_data, test_size=0.2, random_state=42)

X_train_with_performance_data.columns

# Create a StandardScaler to scale the features
scaler = StandardScaler()

# Fit the scaler on the training data and transform both the training and test data
X_train_scaled_with_performance_data = scaler.fit_transform(X_train_with_performance_data)
X_test_scaled_with_performance_data = scaler.transform(X_test_with_performance_data)

with open("scaler_model.pkl", "wb") as f:
    pickle.dump(scaler, f)

# Create a Linear Regression model
# model = LinearRegression()

# Create a Random Forest Regression model
# model = RandomForestRegressor(n_estimators=500, max_depth=3, random_state=42)

# # Create an XGBoost Regressor model
model = xgb.XGBRegressor(
    n_estimators=500,          # Number of boosting rounds (trees)
    learning_rate=0.1,         # Step size shrinkage to prevent overfitting
    max_depth=7,               # Maximum depth of individual trees
    random_state=42
)

# Create an SVR model
# model = SVR(kernel='rbf', C=20)

# Fit the model on the training data
model.fit(X_train_scaled_with_performance_data, y_train_with_performance_data)

with open("regression_model.pkl", "wb") as f:
    pickle.dump(model, f)

# Make predictions on the test data
y_pred_with_performance_data = model.predict(X_test_scaled_with_performance_data)

# Evaluate the model
mse = mean_squared_error(y_test_with_performance_data, y_pred_with_performance_data)
r2 = r2_score(y_test_with_performance_data, y_pred_with_performance_data)
mae = mean_absolute_error(y_test_with_performance_data, y_pred_with_performance_data)

# Calculate Adjusted R-squared
n = X_test_scaled_with_performance_data.shape[0]  # Number of samples
p = X_test_scaled_with_performance_data.shape[1]  # Number of predictors (features)
adjusted_r2 = 1 - (1 - r2) * (n - 1) / (n - p - 1)

# Print the evaluation metrics
print("Mean Squared Error:", mse)
print("R-squared:", r2)
print("Adjusted R-squared:", adjusted_r2)
print("Mean Absolute Error:", mae)

# Assuming 'X' contains your independent variables (excluding the unique identifier) for 'without_performance_data'
X_without_performance_data = without_performance_data.drop(['hurdle_rate_by_dpd30_mean', 'store_number'], axis=1)

# Scale the features using the same scaler used for the 'with_performance_data' dataset
X_scaled_without_performance_data = scaler.transform(X_without_performance_data)

# Make predictions on the 'without_performance_data' dataset
y_pred_without_performance_data = model.predict(X_scaled_without_performance_data)

# Add the predictions to the 'without_performance_data' DataFrame
without_performance_data['hurdle_rate_by_dpd30_mean'] = y_pred_without_performance_data

without_performance_data.head()

without_performance_data.shape

without_performance_data['hurdle_rate_by_dpd30_mean'].describe()

without_performance_data[without_performance_data['hurdle_rate_by_dpd30_mean'] >= 1.03].shape
